requirements.txt
flask==3.0.3
wakeonlan==3.1.0
python-dotenv==1.0.1
